- Updated Domain models - Created and applied Initial migration - Updated Dtos models

+++ Library.Application/Dtos/Book/GetAllBooksDto.cs
@@ -8,8 +8,7 @@ namespace MyLibrary.Application.Dtos.Book
 {
     public class GetAllBooksDto
     {
-
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string Title { get; set; }
         public string AuthorName { get; set; }
         public string Image { get; set; }
